fix: split components per functionality
