+++ app/components/packages/dependencies/Dependencies.js
@@ -7,7 +7,8 @@ import Typography from '@material-ui/core/Typography';
 import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
 import ListItemText from '@material-ui/core/ListItemText';
-import styles from './styles/dependencies';
+
+import styles from './styles';
 
 const Dependencies = ({ classes, data }) => (
   <Paper className={classes.paper}>
