+++ app/components/packages/dependencies/styles.js
