+++ app/components/packages/info/styles.js
